api/Librairie/api/serializers.py
from rest_framework import serializers
from .models import Etudiant, Livre, Personnel, Categorie, Emprunt, Appartient, Auteur

class EtudiantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Etudiant
        fields = '__all__'  # Inclut tous les champs du modèle

class LivreSerializer(serializers.ModelSerializer):
    class Meta:
        model = Livre
        fields = '__all__'

class PersonnelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Personnel
        fields = '__all__'

class CategorieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Categorie
        fields = '__all__'

class EmpruntSerializer(serializers.ModelSerializer):
    class Meta:
        model = Emprunt
        fields = '__all__'

class AppartientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Appartient
        fields = '__all__'


class AuteurSerializer(serializers.ModelSerializer):
    class Meta:
        model = Auteur
        fields = '__all__'